EmployeeAction reducer started
