+++ src/actions/EmployeeActions.js
@@ -1,0 +1,11 @@
+import {
+  EMPLOYEE_UPDATE
+} from './types';
+
+export const employeeUpdate = ({prop, value}) => {
+  return {
+    type: EMPLOYEE_UPDATE,
+    payload: {prop, value}
+  };
+
+};
