+++ src/actions/types.js
@@ -3,3 +3,5 @@ export const PASSWORD_CHANGED = 'password_changed';
 export const LOGIN_USER_SUCCESS = 'login_user_success';
 export const LOGIN_USER_FAIL = 'login_user_fail';
 export const LOGIN_USER = 'login_user';
+
+export const EMPLOYEE_UPDATE = 'employee_update';
